Update README.md
Fixed certain paths for the wget example for a Linux install, and fixed the path for the "how to start automatically in Linux" section. Naming schema hadn't been updated, so the WGET example was 404ing, and the service was not running properly (since the `ExecStart=/var/lib/airconnect/aircast-linux-arm -l 1000:2000 -Z -x /var/lib/airconnect/airupnp.xml` referenced the wrong filename for the exec).